Copy context in WithParam and WithTimeout
