+++ db_no_context.go
@@ -3,6 +3,8 @@
 package pg
 
 import (
+	"time"
+
 	"github.com/go-pg/pg/internal/pool"
 	"github.com/go-pg/pg/orm"
 )
@@ -16,4 +18,30 @@ type DB struct {
 	fmter orm.Formatter
 
 	queryProcessedHooks []queryProcessedHook
+}
+
+// WithTimeout returns a DB that uses d as the read/write timeout.
+func (db *DB) WithTimeout(d time.Duration) *DB {
+	newopt := *db.opt
+	newopt.ReadTimeout = d
+	newopt.WriteTimeout = d
+
+	return &DB{
+		opt:   &newopt,
+		pool:  db.pool,
+		fmter: db.fmter,
+
+		queryProcessedHooks: copyQueryProcessedHooks(db.queryProcessedHooks),
+	}
+}
+
+// WithParam returns a DB that replaces the param with the value in queries.
+func (db *DB) WithParam(param string, value interface{}) *DB {
+	return &DB{
+		opt:   db.opt,
+		pool:  db.pool,
+		fmter: db.fmter.WithParam(param, value),
+
+		queryProcessedHooks: copyQueryProcessedHooks(db.queryProcessedHooks),
+	}
 }
